pyproject.toml
[tool.poetry]
name = "transcriber"
version = "0.1.0"
description = ""
authors = ["Peter Ciccolo <peter.ciccolo@neon.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ffmpeg-python = "^0.2.0"
openai = "^1.3.0"
pysub-parser = "^1.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
